+++ spec/lib/integral/widgets/recent_posts_spec.rb
@@ -3,7 +3,7 @@ require "rails_helper"
 module Integral
   module Widgets
     describe RecentPosts do
-      let(:posts) { Integral::Post.published.limit(2) }
+      let(:posts) { Integral::Post.published.limit(2).order(published_at: :desc) }
       let(:render_options) { {
         partial: 'integral/posts/collection',
         locals: {
@@ -23,17 +23,17 @@ module Integral
 
         context 'amount option' do
           context 'when amount is passed in' do
-            let(:posts) { Integral::Post.published.limit(3) }
+            let(:posts) { Integral::Post.published.limit(3).order(published_at: :desc) }
 
             it 'renders with set amount of posts' do
-              expect(ApplicationController).to receive(:render).with(render_options)
+              expect(Integral::ApplicationController).to receive(:render).with(render_options)
               described_class.render(amount: 3)
             end
           end
 
           context 'when amount is not passed in' do
             it 'renders with default set amount of posts' do
-              expect(ApplicationController).to receive(:render).with(render_options)
+              expect(Integral::ApplicationController).to receive(:render).with(render_options)
               described_class.render
             end
           end
@@ -41,17 +41,17 @@ module Integral
 
         context 'tagged option' do
           context 'when tagged is passed in' do
-            let(:posts) { Integral::Post.published.tagged_with('foo bar'.split).limit(2) }
+            let(:posts) { Integral::Post.published.tagged_with('foo bar'.split).limit(2).order(published_at: :desc) }
 
             it 'renders with set tagged posts' do
-              expect(ApplicationController).to receive(:render).with(render_options)
+              expect(Integral::ApplicationController).to receive(:render).with(render_options)
               described_class.render(tagged: 'foo bar')
             end
           end
 
           context 'when tagged is not passed in' do
             it 'renders with all recent posts' do
-              expect(ApplicationController).to receive(:render).with(render_options)
+              expect(Integral::ApplicationController).to receive(:render).with(render_options)
               described_class.render
             end
           end