+++ CHANGELOG.md
@@ -19,6 +19,7 @@ recommendations of [keepachangelog.com](http://keepachangelog.com/).
 - Moved list edit form to /edit route & removed show route
 - Dashboard displays most recent post & page
 - Post data appears in dashboard widgets when blog is enabled
+- Recent Posts widget orders collection by most recently published
 
 ## v1.0.1